fix: add traversal of the ["compatible"] list
In older versions of dtc, compatible is not serialized,
we need to use split("\x00") to split the first element as a string.

The new version of dtc will automatically serialize compatible.

So now iterate over each element in ["compatible"] list and split("\x00")
to compatible with both old and new versions of dtc

Signed-off-by: Jiali Chen <chenjiali@radxa.com>